+++ Learning/Learning/Learning/Models/QAMarks.cs
@@ -5,7 +5,7 @@
         public Guid ID { get; set; }
         public Guid RegisterID { get; set; }
         public Register Register { get; set; }
-        public int QMarks { get; set; }
+        public float QMarks { get; set; }
         public int TotalMarks { get; set; }
         public string course { get; set; }
     }
